ln-train-batch/requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.26.4
boto3==1.26.163
python-dotenv==1.0.0
pytest==7.4.0
marshmallow==3.19.0